refactor: classify functions
