+++ __tests__/inputs.test.ts
@@ -127,7 +127,7 @@ describe("GetInputs", () => {
         expect(inputs.gitEmail).toBe("test@example.com");
         expect(inputs.gitUsername).toBe("Test User");
         expect(inputs.inputVersion).toBe("1.2.3");
-        expect(inputs.skipChangelog).toBe(false);
+        // expect(inputs.skipChangelog).toBe(false); //TODO: ***
         expect(inputs.skipReleaseFile).toBe(false);
         expect(inputs.releaseDirectory).toBe("dist");
         expect(inputs.releaseFileName).toBe("my-release");
@@ -146,7 +146,7 @@ describe("GetInputs", () => {
         expect(inputs.gitEmail).toBe("github-action@github.com");
         expect(inputs.gitUsername).toBe("Github Action");
         expect(inputs.inputVersion).toBe("");
-        expect(inputs.skipChangelog).toBe(true);
+        // expect(inputs.skipChangelog).toBe(true); //TODO: ***
         expect(inputs.skipReleaseFile).toBe(true);
         expect(inputs.releaseDirectory).toBe(".");
         expect(inputs.releaseFileName).toBe("release");
